Create GetAllResourceStringsForEveryLanguageOnTheSite.v13.sql
This sql script gets all resource strings, for all languages.  For each resource key, it creates a record (empty or not) with the valid localization code, resource key, and value.  The result is a site with 6 languages will have 6 records for every localization key, even if there are no values for all keys in every language.   

This is useful when sending the keys and values to the client to review what needs to be localized.